ai_service/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pytesseract
scikit-learn
numpy
python-dateutil
